Add page for  vacation to Cannes (#4)
* Rider crap

* Update SDK and Sutil package versions

The global.json file was updated to use version 7.0.203 of the SDK with rollForward set to latestMinor. The App.fsproj file was also updated to use version 2.0.10 of the Sutil package instead of version 2.0.9.

* Add Cannes page and link to navbar

This commit adds a new page called "Cannes" and links it to the navigation bar. The new page contains information about must-see locations in Cannes, including descriptions, distances from the port, and website links.

* Add some locations and a stub itinerary