Update eslint (#64)
* Updated Node version

The Node version specified in the .nvmrc file has been updated from 20.14.0 to 21.7.3 for better compatibility and performance improvements.

* Updating eslint to v9

* Cleaning up eslint config a little

* Added linting step to CI workflow

A new step has been added to the continuous integration (CI) workflow. Now, in addition to building the project, the linter will also be run. This ensures that all code adheres to our style guidelines before it is deployed.

* Fix eslint errors and add script